checkpoint: moves ".../Fetching/Error.ts" to ".../Fetching/Error/Request.ts"
